ha-websocket/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
websocket-client==1.8.0
    # via -r requirements.in
